Moving login message into LoginForm component Adding a logic to toggle login form depending on authenticated state in LoginForm function Adding renderLoginForm in initialState.js
