+++ src/components/LoginForm.jsx
@@ -1,9 +1,13 @@
 import React from 'react'
 import { authenticate } from "../modules/authenticate";
-import { useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 
 const LoginForm = () => {
   const dispatch = useDispatch();
+  const authenticated = useSelector(state => state.authenticated)
+  const credentials = useSelector(state => state.credentials)
+  const renderLoginForm = useSelector(state => state.renderLoginForm)
+
   const authenticateUser = async (e) => {
     e.preventDefault()
     let response = await authenticate(e.target.email.value, e.target.password.value)
@@ -14,23 +18,39 @@ const LoginForm = () => {
     }
   }
   return (
-    <form onSubmit={authenticateUser}>
-      <input
-        type="email"
-        name="email"
-        data-cy="input-email"
-      />
-      <input
-        type="password"
-        name="password"
-        data-cy="input-password"
-      />
-      <input
-        type="submit"
-        value="Login"
-        data-cy="btn-login"
-      />
-    </form>
+    <>
+      {renderLoginForm &&
+        <form onSubmit={authenticateUser}>
+          <input
+            type="email"
+            name="email"
+            data-cy="input-email"
+          />
+          <input
+            type="password"
+            name="password"
+            data-cy="input-password"
+          />
+          <input
+            type="submit"
+            value="Login"
+            data-cy="btn-login"
+          />
+        </form >
+      }
+
+      <p data-cy="login-message">
+        {authenticated === false ?
+          "Invalid credentials. Please confirm your email and password."
+          :
+          (authenticated === true ?
+            `Hello, ${credentials.uid}!`
+            :
+            ""
+          )
+        }
+      </p>
+    </>
   )
 }
